combined Vlans services into Vlans_Services.py, modified validate Vlans function
